Fixed the commuteBinALMD peephole optimzation.
I was only checking that the operation (i.e., +, *, AND, OR, XOR)
allowed it's operands to be commuted, not that the operands could be
commuted without changing the meaning of the program.  In the good case,

mov 2,Y  ==>  mov X,Y
add X,Y       add 2,Y

We only commute if the first src is an immediate and the second source
is not.  (The idea being, if Y is a loop variable, then presumably
there was some move of Y to X previously, now there is a chance that
we'll be able to elimate the mov X,Y or at the very least, try to get
them into the same registers.)

But, in the bad case:

mov 2,Y  ==>  mov Y,Y
add Y,Y       add 2,Y

Strangely, this only came up from something in the machine IL like:

RI(3) = Int_mulCheck [2147483647, 2]  Overflow L_51

which became

mov 2^31-1,ov ==> mov 2^31-1,ov ==> mov ov,ov     ==> add 2^31-1,ov
imul 2,ov         add ov,ov         add 2^31-1,ov

Now ov was live through the whole program and had some random value;
Actually, I suspect that it had the value 0, because the program went
into an infinite loop.