Added -native-cutoff flag for live stack slots.
This flag limits the values carried in registers across blocks.  Only
values with a use < cutoff instructions down some execution path are
carried in a register.  This prevents long lived values from being
carried in registers far from their use.  Currently defaults to 100,
but I'll need to play a little more with it to get a better default.