Previously, after the last C-call in a block, we would see:
	call GC_gc
	addl $16,%esp
	movl %esp,c_stackP

But, there is no need to store the value back in c_stackP; it hasn't
really changed.  And, that being the case, there is no need to pop the
stack after the return from a C-call.  So, we recognize this pattern
don't produce extraneous code.

Note, we do need to adjust the %esp after C-calls that are not the
last in the block; otherwise we risk exhausting the C-stack.