An abortive attempt to upgrade to a newer SML/NJ.  I tried 110.33 and 110.34, but both of those exhibited compiler bugs when compiling machine.fun.  I tried 110.32, but that caused a very strange link-time overflow exception, although it claimed to successfully compile everything.  I tried to track down the link-time overflow, but was unable to because it depended on the presence of print statements in the offending file.  So, I conclude there was actually something much buggier going on in the SML/NJ compiler.
In any case, we will stick with 110.9.1 for now.  I made a bunch of changes
(mostly changing Group to Library in CM files) that should make it easy to
upgrade in the future, should a suitable version of SML/NJ become available.

Also, there was some bizareness with the 110.9.1 CM that caused me to add a
mlton/codegen/x86-codegen/word.sml, which simply redefines the Word structure.
If that file is not there, then CM gets confused and uses the basis library Word
instead of the MLton library Word -- I have no idea why and I have no idea why
the fix works.