Modified ProfileInfo to ensure that level 0 is a cps name, level 1 is a label found in .cps, and level 2 is a labeled block in the assembly.
Note, you can still end up with L_1,L_2 for level 2 in mlprof, because something like:

fib_0:
statementLimitCheckLoop_1:

ends up with two profile labels that have the same level 0 and level 1
fields, but different level 2 fields.  But, the also have the same
assembly address, so I combine the sets of fields.  (I guess we could
fix this by also keeping track of the unique id attatched to the
profile labels.  Then we just take the later of the two.)