Removing dependencies on $(NAME).cm.  Must manually do make mlton.cm if a soucres.cm file changed.
Rationale:

We decided that /front-end/ml.grm.* and /front-end/ml.lex.sml were
derived files, and removed them from the repository.  Due to the way
CM and cmcat work (failure if a file in a sources.cm does not exist),
we made mlton.cm depend on these derived files.  This has the side
effect of making mlton.cm out of date when starting from clean
sources.  Anyways, I mention this because the argument is that
mlton.cm is part of the sources, because someone without cmcat can't
make it.  But, if the default trajectory of the Makefile is going to
force rebuilding of mlton.cm using cmcat, that seems contradictory.