Modified the benchmark program to allow any number of mlton compilers to be specified on the command line.  Invoke it like:
benchmark -mlton "/usr/local/bin/mlton"  -mlton "mlton -new-flag true" \
          -mlton "mlton -new-flag false"  bench1 bench2 ...

It automatically splits the argument to -mlton at #" "'s and uses the
first value as the compiler and the remainder as arguments.  Works out
fine with Process.call'.

I left in the other compilers as command line options as well.  The first
compiler specified on the command line will be used as the base for
normalizing run-times.

The only downside to this right now is that the headers for the resulting
table inherit the whole compiler string as the name (so it's a bit wide).

For really serious benchmarking, I was thinking of adding a command line
option like:

-flag "mlton;;-new-flag1,{true,false};-new-flag2,{0,1,2}"

that would specifiy a compiler, and a sequence of options and option
arguments and the benchmark program would automatically produce all
combinations of flags.  But, I can get this effect with the above
modification and a little cutting and pasting in the Makefile.