Fixed an inner/outer bug:
fun main_0 ()
  = let
       ...
       fun x_6 (env_14)
         = let
              ...
              fun L_171 (env_13) = x_6 (x_258 (env_3))
              ...
           in
              ...
           end
       fun x_188 (env_10) = x_6 (x_258 (env_3))
       ...
    in
       ...
    end

The transformation places the contified x_258 immediately after the
declaration of x_6, out of the scope of the transformed L_171 body.

Solution: detect this case and mark the function as uncontifiable

The solution is pretty simple; change rule 5 in the definition of the
graph G from
   {(k,g) | (f,g,k) \in N and R(f)}
to
   {(k,g) | (f,g,k) \in N and R(f) and (f,g,k) \not\in body(k)}
 U {(Root,g) | (f,g,k) \in N and R(f) and (f,g,k) \in body(k)}

That is, when there is a non-tail call to g with continuation k in the
body of continuation k, then make Root a predecessor of g in the graph G;
this will guarantee that idom(g) = Root and Adom(g) = Unknown.

This won't hurt us that much; things that were going to get contified at k
because they were tail-called from g will just get contified in g, because
g will still dominate them in G.