Attempt to solve the following problem that we've had forever with mlton/Makefile and mlyacc/Makefile.  The problem comes from the following statements, which are conflicting
1. we don't want to depend on others having cmcat.
2. cmcat generates mlyacc.cm from sources.cm.
3. mlyacc.cm is generated, and should not be part of the source.
4. yacc.lex.sml is generated, and should not be part of the source.
5. sources.cm mentions yacc.lex.sml.
6. cmcat failes if any of the files in sources.cm do not exist.

(1), (2), and (3) obviously conflict, because if we don't include
mlyacc.cm in source, then whoever does a make must have cmcat.  We
decided to resolve this conflict in favor of (1), and include
mlyacc.cm in the source.

(1), (4), (5), and (6) also conflict, in a nontrivial way.  Because of
(5) and (6), the Makefile expresses the dependence of mlyacc.cm on
yacc.lex.sml.  Because of (4), if someone does a clean make, then
yacc.lex.sml generated.  Hence, if someone does a clean make,
mlyacc.cm needs to be remade, which conflicts with (1).  We
(mistakenly, IMO) resolved this conflict in favor of (4), (5), and
(6).

It's silly for us to choose one way in the first conflict and the
other in the second conflict.  I still think (1) is worthwhile, and
would like to resolve the second conflict in its favor.  We could
either drop (4) or the Makefile dependence that follows from (5) and
(6).

To resolve the problem, Matthew suggested this approach, which keeps the
dependency on generated files unstated, but enforces it before getting to cmcat
which will require their existence.