Change to handling of bug continuations.
The current backend (and now SSA type-checker) assumes that live
variables down a handler are live at the top of the corresponding
continuation.  This causes problems when we introduced a bug
continuation at non-tail calls with different handlers.

This might go away in the future, and the extra wrapper added by the backend.