Minor "bug" in initialization code. When not restoring a world, GC_init initializes currenThread to an empty stack.  This places stackTop pointing to the word below stackBottom.  So, we've really been stomping on currentThread->stack->used when doing
               *(pointer *)gcState.stackTop = (pointer)&ml;            \

This never causes a problem, because upon reentry to the runtime, we
update currentThread->stack->used with the new stackTop.  But, we
probably shouldn't rely on being able to stomp on the word below
stackBottom.  Hence, the fix.

This was revealed when I tried moving exnStack from GC_thread to GC_stack.