Minor changes to the setting of profileInfo in backend.
The profiling info displayed in -d 1 should always be labels that appear
in SSA.  This eliminates the problem with continuation stubs, handler
stubs, and case dispatch stubs getting Machine IL names.  Continuation
stubs' and handler stubs' ticks get allocated with the "real" continuation
or handler block.  Cases are a little trickier.  The final stubs that go
to the label mentioned in SSA are associated with the SSA label.  You can
also have additional stubs that split on int/pointer, etc.; these stubs'
ticks get allocated to the SSA block that has the case.