Port of simplifyTypes and useless from CPS to SSA.
The fact that Transfer.Return is of Var.t instead of Var.t vector is
forcing useless to keep more things than necessary useful; in
particular, in 1.sml, useless determines that all raises are usless
and hence all handler args are useless, but we can't rewrite raises to
raise nothing.  So, intead, we force the exnVal to be useful, which in
turn makes all raised values useful (as opposed to being replaced by
bogus values).  Probably not that big a deal -- how many programs will
really end up with useless raises.