Fixed bug that caused an unclosed string error message on "abc\5xyz".  As Henry said, the problem was that {ws} matches the empty string, hence the lexer switches to from the <S> state to the <F> state.  Then in the <F> state only the final rule works.
The fix was simply to change {ws} to {nrws}, which requires at least one
whitespace character.  This is correct by the rule on page 4 of the Definition
that says you must have at least 1 `formatting characters' to go into this
state.

Now, MLton more sensibly reports an illegal string escape.