Removing as much of the CPS simplifier as possible.
Attempting, but failing, to clean up the property list problems with the SSA
simplifier.  The checked in compiler successfully passes all regressions and
benchmarks, but fails during self-compilation due to a property list numLinks
overflow.  Something really bad is happening, because the maximum property list
length is 7577, instead of the usual 5 or so with CPS simplification.