Fixed a bug is isStackHeader, which would mistakenly return true for arrays with at least 2^28 elements.
Renamed isStackHeader to isStackOrContHeader to indicate that it should return
true for stacks or conts.

Fixed a bug in GC_MAX_ARRAY_LENGTH.  It was too high.  If an array has >= 2^29
elements (i.e. >= 0x20000000 elements), then its length will have a tag bit set,
and the length will therefore be indistinguishable from a header word.  This
would be bad, since the gc forwarding loop that walks through the queue in
toSpace, and in particular the toData function, must be able to distinguish an
array length from an object header.