Working on a new x86-generate-transfers.fun.  It's not quite there yet, but you can test it out with -native-new-gen-transfer true.
I essentially want a better block ordering, as well as getting better
alignment for loop headers.  What seems to be screwing up is jo
instructions -- I want to push the overflow case "away", but doing so
can increase the size of the immediate label in to jo instruction,
making loops larger (more expensive) than they would otherwise be.  I
did fix it so that we always fall through to the skipGC case of a
limit check, which should get the right static branch prediction each
time.