Removed the last vestiges of CPS.
This involved a complete rewrite of Ssa.DirectExp, which now uses an, er..,
direct representation of terms.  The linearize functions are more efficient
because they can cons onto a list of blocks.  They also use as an intermediate
result a list of statements and a transfer instead of a block label.  This
avoids the creation of lots of small blocks.

Some minor tweaks to closure convert were necessary.

I also fixed a shrinker bug that showed up, having to do with simplifying case
expressions.

I am pleased to that the SSA pre codegen is as fast as the CPS pre codegen in
20011006.  Here is the log for compiling this new version of MLton with
20011006.

MLton starting
   Compile SML starting
      pre codegen starting
	 gcc -DNODEBUG -DMLton_safe=TRUE -DExn_keepHistory=FALSE \
	     -DMLton_detectOverflow=TRUE -I/home/sweeks/mlton/include -o \
	     /tmp/fileDwp23g /tmp/filezSjEDy.c -L/home/sweeks/mlton/lib \
	     -lmlton -lm -lgmp
	 /tmp/fileDwp23g /tmp/file1k14FW
      pre codegen finished in 101.11 + 115.58 (53% GC)
      x86 code gen starting
      x86 code gen finished in 161.84 + 56.67 (26% GC)
   Compile SML finished in 262.95 + 172.25 (40% GC)

And here is the log of a self-compile.

MLton starting
   Compile SML starting
      pre codegen starting
	 gcc -DNODEBUG -DMLton_safe=TRUE -DExn_keepHistory=FALSE \
	     -DMLton_detectOverflow=TRUE -I/home/sweeks/mlton/include -o \
	     /tmp/file23g1k1 /tmp/fileEDyDwp.c -L/home/sweeks/mlton/lib \
	     -lmlton -lm -lgmp
	 /tmp/file23g1k1 /tmp/file4FWLIB
      pre codegen finished in 110.95 + 103.99 (48% GC)
      x86 code gen starting
      x86 code gen finished in 579.68 + 175.36 (23% GC)
   Compile SML finished in 690.63 + 279.35 (29% GC)

The pre codegen times are almost identical.

The dramatic increase in x86codegen time is due to the limit check problem that
has not been fixed, and a recent performance problem introduced in
directed-graph loopHeaders.  Hopefully, the latter will be fixed imminently.