Eliminated the distinction between Thread_switchTo and Thread_switchToCont. Now, there is only Thread_switchTo, and it is the presence of this primitive that means that a program uses threads/continuations.  So, globalization, constant propagation, and localRef only need to look for this primitive.
Changed Thread_copyShrink t to Thread_copyCurrent ().  The only uses of
Thread_copyShrink were in the basis library thread.sml and cont.sml, which
always called Thread_copyShrink with the current thread.  This is now hardwired
into Thread_copyCurrent.  Also, Thread_copyCurrent is the only place where
control-flow is not explicit.  That is, any use of Thread_switchTo, or any place
where the runtime is entered, may jump to a use of Thread_copyCurrent.  So, for
example, globalization says that any code after a Thread_copyCurrent may be
executed more than once.

Fixed a bug in ccodegen.h in the LimitCheck macro.  The bytes must be evaluated
outside the InvokeRuntime macro, which changes the stack pointer.