Extended the SSA type-checker to verify that all case transfers are exhaustive and irredundant.  (All Char/Int/Word8/Word must have a default, even though a Char/Word8 case could be exhaustive.)
src/regression/polymorphic-recursion.sml does not translate into an
SSA program that satisfies these invariants (it has an exhaustive case
with a default).  Extended the XML simplifier to drop default cases on
exhaustive Cases.Con's.  Perhaps this could go in the SSA shrinker,
but the shrinker does not have the datatypes available.