Eliminated max-heap-swap runtime option.
Also, made modifications to the runtime to behave better in situations where the
amount of live data is a signifant fraction of the amount of RAM, based on code
from PolySpace.  MLton executables by default can now use more than the
available amount of RAM.  Executables will still respect the max-heap runtime
arg if it is set.