Reworked object header words so that Array.maxLen = valOf Int.maxInt.
Also fixed a long-standing minor bug in MLton, where Array.array (Array.maxLen,
...) would raise Size instead of attempting to allocate the array.  It was an
off-by-one error in the meaning of Array.maxLen.

This required changing a few regression tests that had assumed that
Array.maxLen + 1 wouldn't overflow.

There is still a bug remaining because the size computations for array limit
checks may overflow, and that is not checked for.  I plan to fix that soon by
moving the computations out of the codegens and into the backend's limit check
pass.