Peform knownCase on all tycons (don't exclude booleans).
I thought this would be prohibitively expensive, but it really wasn't.

I found a couple of cases where this would help in small programs;
more importantly, it can prevent the variable holding the boolean
result from being live past the transfer.  This means that the codegen
can avoid holding onto the intermediate result; that is, we avoid a
setcc and a movzbl.