Improvements to threads/signal handling.
Thread.atomicEnd now checks that canHandle > 0.

Removed the implicit atomicEnd from the end of Thread_switchTo.  The atomicEnd
is now explicit in the thread implementation in the basis library -- whenever a
thread resumes, it performs an atomicEnd to match the atomicBegin that was done
by the switcher.  There is a little bit of trickiness with the threads that are
passed by the runtime system to the signal handler, which were interrupted and
don't know to do the atomicEnd.  So, Thread.fromPrimitive adds it for them.

Removed the implicit Thread_switchTo at the end of GC_finishHandler.  Now,
Thread_finishHandler is a simple FFI call instead of a primitive, and the signal
handler thread switches to the thread returned by the handler.  Unfortunately,
this means that I had to create a new primitive MLton_handlesSignals, which is a
noop, just so that the backend can detect if the program handles signals.