Fixed bug in slower when reserveEsp = true.
There is a hack in the register allocator to _not_ adjust c_stackP
after the last C call in a basic block; the reasoning is after the
final C call, popping all the arguments by adding to %esp will simply
make %esp equal to c_stackP, which we'll just load in the next block
that makes a C call.  So, I was saving one instruction per C call.  In
Cygwin (or with reserveEsp = true) this is very bad, because we'll
never refetch c_stackP from memory, we'll just assume it's up to date
in %esp.  So, with slower.sml, there is a loop that executes for
4304967296 times, each time through the loop making two C calls -- we
simply ran out of stack space.