Implemented World.save without using fork so that it can be reliably used on systems (like Cygwin) that don't have a reliable fork.
Also, changed World.save so that it is not allowed to be used from within a
signal handler.  Instead, you must explicitly use World.saveThread.

GC_saveWorld now operates on a file descriptor instead of a filename.  This way,
the exception can get raised in ML if there are problems with the file.

Renamed the "safe" runtime functions like swrite to sfwrite to more accurately
reflect the underlying call.  Now, swrite is a wrapper around write and sfwrite
is a wrapper around fwrite.