Version of knownCase that fixes the bug exhibited by Anoq's parser program.
One bug wasn't the actual problem, but could have been a problem.
When rewriting a Goto to statements that move actuals to formals and
then peform the destination block's transfer, the moves were done
naively, so at a loop where a formal was also an actual, I could
accidently move an actual into a formal and later move that newly
bound formal (which was also an actual) into another formal.

Anoq's bug was more subtle.  When going around a loop, we need to
update any variable whose value is based on the "old" argument.  This
essentially required maintaining a stack of variable replacements refs
and when going around a loop, copy the old arguments to fresh
variables, update the refs, and then move the new actuals into the
formals (correctly, as described above).