MAIL
This is a major rewrite of the heap resizing code.  The code now closely follows
the idea that there should always be enough space in RAM to fit a semispace plus
the amount of live data.  For details, see the comment before computeSemiSize in
gc.c.  The functions that handle resizing are prepareToSpace, which runs before
a GC and allocates toSpace and resizeHeap, which runs after a GC and may shrink
the newSpace and/or deallocate old space.

I also completely rewrote the code for handling when a GC fails to produce
enough space for bytesRequested.  See the code in GC_doGC for details.  The new
approach involves shifting the live portion of the heap into a semispace that is
just large enough to hold it, reallocating a new semispace of the desired size,
and then shifting the live portion into the new semispace.  This works better
when address space problems crop up.  In doing this, I moved translateHeap from
GC_world.c to gc.c.