Fixed bug in overloading of / so that the following now type checks:
	fun f (x, y) = x + y / y

The problem is that because + and / are not constrained, the type
checker chooses the default type for +, int * int -> int.  It is then
screwed because it can't chose that type for /.  The problem happens
when there are overloaded variables that have some compatible type (in
this case real) but one of whose default types (int) is not a valid
instance of the other.

The fix was to simply not declare / as an overload.