Changed the interface to IntInf_* operations to peform their own allocation.  Instead of passing in a word array, a word corresponding to the maximim number of bytes needed for the result is passed; the limit check insertion pass extracts this word and uses it to ensure that sufficient heap space is available at the time of the call.
For example, under the old system:

[fluet@localhost temp]$ mlton-stable -v3 -drop-pass flatten -drop-pass
inline -drop-pass leafInline -drop-pass localFlatten. -debug ???
-gc-check first -type-check true smith-normal-form.sml

would either assert in IntInf.c or dies with
Out of memory: attempt to allocate more than 2147483647 bytes.
depending on the setting of -debug.  Now, both versions execute
without errors.

Also modified the treament of arrays in RSSA and MACHINE.  There is a
new primitive Array_allocate (int, word, word) which takes the number
of elements, the bytes reserved, and the header word.  This means that
all limit check operations look for Statement.PrimApp's.  It also
means that the ssaToRssa pass computes the full number of bytes
reserved (including header and size words).  This complicates the
implementation of Array_allocate a little in the C backend, but
simplifies it in the native backend.