MAIL mark compact GC and backend/codegen changes
This is the first checkin of the mark compact GC.  It is disabled for now, but
has passed all the regressions and a self compile with the C codegen.  There
were also a number of backend and codegen changes, which I'll try to highlight
below.  I've got everything completely working with the C codegen, but it's
broken with the native codegen for now.  Matthew, if you could take a look, that
would be great.

For the mark compact GC, new header words have been introduced.  For details of
the header word layout, see the comment at the top of gc.h.  A counter word was
added to arrays for use during marking.  To integrate the mark compact with the
GC, I still need to handle heap resizing and decide when to switch between
stop-and-copy and mark-compact.

At the high level, the backend/codegen changes were motivated by moving as much
knowledge as possible from the codegens into the backend, do avoid duplication
across codegens.  The biggest change to the backend is that I eliminated the
Runtime Transfer from Rssa and Machine.  What used to be implemented with
Runtime is now implemented as a CCall.  CCall now has more information about the
function that is being called, including whether it may GC, whether it modifies
the frontier or the stackTop, whether it may return, whether they need bytes
free in the heap, and more.  See backend/c-function.sig.  I also modified the
IntInf_ calls to use the normal C calling convention (although the codegen must
use the fact that they modify the frontier) and modified the backend to
recognize which primitives are directly implementable as CCalls.  The upshot of
these changes is that a lot of primitives do not have to be handled specially by
the codegens anymore.  Instead, the codegens need to implement a more
complicated version of CCall.  As I mentioned above, this all works with the C
codegen, but not yet with the x86 codegen.

I added a new option, -inline-array {true|false}.  When true, arrays are
allocated inline, as they used to be.  When false, they are allocated and
initialized by a C routine, GC_arrayAllocate.  This routine could be used as a
hook in the future to special treatment by the runtime of large or pinned
arrays.  As soon as x86 codegen is working again, I'll run tests and see if
it hurts performance to switch the default to false.