MAIL mark compact GC and x86-codegen changes
Here is a working checkin of the native codegen with the mark compact
GC.  Passes all regressions (with -debug true -gc-check first) and a
self compile, both with MARK_VERIFY=TRUE.  I've left the mark compact
GC disabled for now.

I added an additional check to CFunction.isOK that asserts
maySwitchThreads => returnTy = NONE
Nothing currently violates this check, but I was getting a headache
trying to figure out the right caching of values across a thread
switching function that returns a value, not to mention which thread
should receive the returned value.

The new CFunction interface is nice; it lets me express a (trivial)
optimization of GC calls.  In particular, when a program doesn't
handle signals, then a GC won't ever change threads.  Therefore, we
don't need to do an indirect jump to stackTop on return, since we know
that that will be the same return address that we pushed before the
call.  Overall, I doubt it will make any performance difference, but
it was easily added.

There are a couple of outstanding issues with threads, but nothing
that inhibits compilation.  My straightforward port of ccodegen.h's
Thread_switchTo to native assembly is busted, but peforming
Thread_switchTo via a C-call into Thread.c works fine.  (Really don't
know why.)  There is also some extraneous shuffling of the return
value from a C-call that mayGC; this is a side-effect of having
previously assumed that runtime calls don't return.  The basic issue
is that the way the return value is cached was partially relying on
the fact that a CCall was always followed by a unique CReturn to which
the code generator could just fall thru, thereby never needing to move
the return value anywere.  While a mayGC call that returns will also
have a unique CReturn, we can't fall thru, because the frame layouts
data is prefixed to that code in order to GC the top frame.  So, we
need to do an explicit jump, and I need to go back to
x86-live-transfers.fun to instruct it to cache the returned value in a
register over that jump.  Right now we spill it and reload it, which
is semantically correct, and since it is a write followed by a read,
probably always hits the cache, but it shouldn't be too hard to fix.

Other than that, every thing seems to be in order.  I also just
started a regression with -debug true -gc-check first -inline-array
false, and I didn't get any failures through "list".