MAIL Turned on the mark-compact GC into the runtime.  The collector now decides on a per-collection basis whether to use mark-compact or cheney copying (see doGC in gc.c).  It uses copying if it believes the copy can be performed in memory, otherwise it uses mark-compact.
Added heap resizing code that works with both copying and mark-compact.  There
are lots of hardwired constants that surely need to be tweaked.

Reorganized the code in gc.c quite a bit and cleaned it up.

Moved loadWorld and saveWorld from GC_world.c back to gc.c, which meant that a
lot of GC internals didn't need to be exported in gc.h anymore.