MAIL thread bug fixes
Fixed two bugs with threads/signals.  The first was a bug in how the basis
library used GC_copyCurrentThread, which had returned the copied thread as a
result of the C function.  This worked fine in the thread that made the copy,
but unfortunately, this caused problems when switching to the copy of the
thread, because the copy also assumed that a result would be returned.  Of
course when switching there was no result, so whatever happened to be in %eax
was taken as the result.  The fix was to have GC_copyCurrentThread put its
result in gcState.saved, and only to access this result in the thread that makes
the copy.

The second bug happens when a thread fails a limit check for k bytes due to a
signal, and the signal handler switches to another thread.  The problem is that
when switching back to the original thread, there was never a check for the k
bytes.  To fix this, I put a limit check at the end of GC_switchToThread that
checks to make sure that the bytesFree of the thread being switched to is
available.  I also changed thread switch in the backend so that it
ensuresBytesFree, and so that whenever a thread is switched out, it sets
bytesFree.

Thread switching is now quite a bit slower than it used to be, since it has to
go through a couple of C calls and do a little bit of redundant work.  If we
ever need to, we can speed it up again by inlining GC_switchToThread in the
backend (not the codegens!).