An attempt at a script to check a program under the MLton implementation of the basis library; takes a .cm file (generally mlton,mllex,mlprof,mlyacc) prepends the MLton implementation of the basis suitably massaged and pumps the result through SML/NJ. In theory, you shouldn't need to pass a -stubs.cm file. I don't know how much of SML/NJ's basis ends up polluting the environment in which the program is checked; in theory, the MLton implementation should re-export everything that was in the SML/NJ basis.
I wanted to make sure that the lib/mlton/* files were o.k. relative to
the new basis, particularly in terms of vectors and arrays.  Since
Steve reimplements everything from first principals (i.e., length,
sub, and unfoldi), it's actually fine.  There are lots of conflicts
between types of ints.  I don't know if this is a pollution problem, a
real problem (possibly, because I see some uses of LargeInt and IntInf
as if they were equal, which they aren't under the opaque signatures
of the new basis), an overloading problem (in the sense that SML/NJ
reads an integer literal as it's Int.int, which by now is so hidden by
other int's that it doesn't seem to match any other type in scope),
or a problem with the way I'm building the basis environment.  (Note,
there are some deviations from the spec in terms of opaque signature
matching and where clauses that I'm living with so that check-basis at
least works.)