MAIL card maps and machine IL semantics change
Added code to the runtime to allocate space for card maps.  Added code to
backend/ssa-to-rssa so that the mutator does cardmap updates when writing to
pointers arrays or pointer refs.

All regressions pass with the C codegen, but the native codegen is currently
broken, because I changed the semantics of ArrayOffset so that it does not do
the derefence.  I needed to do this because the address in an ArrayOffset is
treated as a value for indexing into the cardmap.  The relevant change in the C
codegen is that

#define ArrayOffset(ty, b, i) (*(ty*)((b) + ((i) * sizeof(ty))))

is changed to

#define ArrayOffset(ty, b, i) ((b) + ((i) * sizeof(ty)))

Matthew, can you please make the relevant change to the native codegen.
Thanks.